Commit before merge

+++ SOPS/Models/ProductRating.cs
@@ -11,7 +11,8 @@ namespace SOPS.Models
     {
         [Key]
         public int Id { get; set; }
-        public float Rating { get; set; }
+        public int Rating { get; set; }
+
         public string UserId { get; set; }
         public virtual ApplicationUser User { get; set; }
